Split opts out into separate file
